Updated build script

+++ build.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
-rm -rf build
-mkdir build && cd build
-# cmake .. -DCMAKE_BUILD_TYPE=Debug
-cmake .. -DCMAKE_BUILD_TYPE=Debug -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_C_COMPILER=clang
+clear && rm -rf build && mkdir build && cd build
+
+# gcc (debug)
+cmake .. -DCMAKE_BUILD_TYPE=Debug
+
+# clang (Debug)
+# cmake .. -DCMAKE_BUILD_TYPE=Debug -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_C_COMPILER=clang
+
 make -j
 # strace ./build/Debug/vulkan-sample
